refactor(web): split LegendSection, fix weapon time-to-kill division by zero
